pyproject.toml
[build-system]
requires = ["setuptools==63.4.1"]
build-backend = "setuptools.build_meta"

[project]
name = "capabilities"
version = "0.0.8"
authors = [
  { name="Blazon AI", email="support@blazon.ai" },
]
description = ""
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"
dependencies = [
  "fire",
  "requests",
  "dacite",
  "asyncio",
  "aiohttp",
  "pdfminer.six",
  "tiktoken",
  "blobfile",
  "pydantic",
  "pydantic[dotenv]",
  "termcolor",
  "rich",
]

[project.optional-dependencies]
search = [
  "pyllamacpp",
  "python-magic",
  "bs4",
  "diskcache",
]
nomic = [
  "nomic",
]
sentence-transformers = [
  "sentence-transformers"
]

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "pytest-snapshot",
  "hypothesis",
  "numpy",
  "torch",
  "nomic",
  "sentence-transformers",
  "requests",
  "dacite",
  "aiohttp",
  "tiktoken",
  "pyllamacpp",
  "python-magic-bin",
  "bs4",
  "diskcache",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]
[[tool.hatch.envs.all.matrix]]
python = ["3.7", "3.8", "3.9", "3.10", "3.11"]
